+++ src/frontend-pdoc-commons/components/pdoc-list-item/pdoc-list-item.component.spec.ts
@@ -2,7 +2,6 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {NO_ERRORS_SCHEMA} from '@angular/core';
 import {PDocListItemComponent} from './pdoc-list-item.component';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {Router} from '@angular/router';
 import {TranslateModule} from '@ngx-translate/core';
 import {PDocDataServiceStub} from '../../../testing/pdoc-dataservice-stubs';
@@ -23,7 +22,7 @@ describe('PDocListItemComponent', () => {
                 { provide: GenericAppService, useValue: new AppServiceStub() }
             ],
             schemas: [NO_ERRORS_SCHEMA],
-            imports: [NgbModule.forRoot(),
+            imports: [
                 TranslateModule.forRoot(),
                 AngularCommonsModule]
         })